requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf 
langchain 
langchain-nvidia-ai-endpoints
faiss-cpu 
sentence-transformers 
transformers>=4.30.0 
requests 
rank_bm25 
numpy 
tqdm 
python-dotenv 
orjson